+++ packages/views/src/useBinding.tsx
@@ -1,5 +1,6 @@
 import { BindingProperty, ensureObservableProperty, PropertyName, PropertyType } from "@frui.ts/helpers";
 import { action, get, isObservable, isObservableMap, isObservableProp } from "mobx";
+import { IBindingProps } from "./bindingProps";
 
 export function getValue<TTarget, TProperty extends PropertyName<TTarget>>(
   target: TTarget | undefined,
@@ -72,25 +73,12 @@ export function setValue<TTarget, TProperty extends BindingProperty<TTarget>>(
   }
 }
 
-export function useBinding<TTarget, TProperty extends PropertyName<TTarget>>(
-  target: TTarget | undefined,
-  property: TProperty | undefined
-): readonly [TTarget[TProperty], (value: TTarget[TProperty]) => void];
-
-export function useBinding<TKey, TValue, TTarget extends Map<TKey, TValue>>(
-  target: TTarget | undefined,
-  key: TKey | undefined
-): readonly [TValue, (value: TValue) => void];
-export function useBinding<TTarget, TProperty extends BindingProperty<TTarget>>(
-  target: TTarget | undefined,
-  property: TProperty | undefined
-): readonly [PropertyType<TTarget, TProperty>, (value: PropertyType<TTarget, TProperty>) => void];
-
-export function useBinding<TTarget, TProperty extends BindingProperty<TTarget>>(
-  target: TTarget | undefined,
-  property: TProperty | undefined
-) {
-  const value = getValue(target as any, property);
-  const setter = (value: PropertyType<TTarget, TProperty>) => setValue(target as any, property, value);
+export function useBinding<
+  TTarget,
+  TProperty extends BindingProperty<TTarget> = BindingProperty<TTarget>,
+  TValue = PropertyType<TTarget, TProperty>
+>(props: IBindingProps<TTarget, TProperty, TValue>) {
+  const value = getValue(props.target, props.property) as TValue;
+  const setter = (value: TValue) => setValue(props.target, props.property, value);
   return [value, setter] as const;
 }