+++ packages/views/src/view.tsx
@@ -6,18 +6,22 @@ interface ViewProps {
   vm: any;
   context?: string;
   useLifecycle?: boolean;
-  fallbackMode?: "message" | "empty";
+  fallbackMode?: "message" | "children";
 }
 
-const View: React.FunctionComponent<ViewProps> = ({ vm, context, useLifecycle, fallbackMode }) => {
+const View: React.FunctionComponent<ViewProps> = ({ vm, context, useLifecycle, fallbackMode, children }) => {
   if (!vm) {
-    return <React.Fragment />;
+    return <React.Fragment>{children}</React.Fragment>;
   }
 
   const FoundView = fallbackMode ? tryGetView(vm.constructor, context) : getView(vm.constructor, context);
 
   if (!FoundView) {
-    return fallbackMode === "message" ? <span>Could not find a view for {vm.constructor.name}</span> : null;
+    return fallbackMode === "message" ? (
+      <span>Could not find a view for {vm.constructor.name}</span>
+    ) : (
+      <React.Fragment>{children}</React.Fragment>
+    );
   }
 
   if (!!useLifecycle) {